rename tierSelector
